Add VIP Big Point functionality with three-day sign-in support and related DTOs
