+++ bruno/api.bilibili.com/x/vip/folder.bru
@@ -1,0 +1,3 @@
+meta {
+  name: vip
+}
