+++ bruno/api.bilibili.com/x/vip/vip_center/sign_in/folder.bru
@@ -1,0 +1,3 @@
+meta {
+  name: sign_in
+}
