+++ bruno/api.bilibili.com/x/vip/web/folder.bru
@@ -1,0 +1,3 @@
+meta {
+  name: web
+}
