+++ bruno/api.bilibili.com/x/vip/web/vip_center/folder.bru
@@ -1,0 +1,3 @@
+meta {
+  name: vip_center
+}
