+++ bruno/big.bilibili.com/x/vip/privilege/folder.bru
@@ -1,0 +1,3 @@
+meta {
+  name: privilege
+}
