+++ src/Ray.BiliBiliTool.Agent/BiliBiliAgent/Dtos/BaseAppRequest.cs
@@ -7,24 +7,26 @@ public class BaseAppRequest
 
     public string build { get; } = "7720200";
 
-    public string c_locale { get; } = "zh_CN";
-
-    public string channel { get; } = Constants.Channel;
-
     public int disable_rcmd { get; } = 0;
-
-    public string from_spmid { get; } = "united.player-video-detail.player.continue";
 
     public string mobi_app { get; } = "android";
 
     public string platform { get; } = "android";
 
-    public string s_locale { get; } = "zh_CN";
-
     public string statistics { get; } =
-        "{\"appId\":1,\"platform\":3,\"version\":\"7.72.0\",\"abtest\":\"\"}";
+        "{\"appId\":1,\"platform\":3,\"version\":\"8.45.1\",\"abtest\":\"\"}";
 
-    // public long ts { get; set; }
+    /// <summary>
+    /// 当前时间（毫秒）
+    /// </summary>
+    /// <sample>1748445354567</sample>
+    public long t { get; set; } = DateTimeOffset.UtcNow.ToUnixTimeMilliseconds();
+
+    /// <summary>
+    /// 当前时间（秒）
+    /// </summary>
+    /// <sample>1748445354</sample>
+    public long ts => t / 1000;
 
     public string? sign { get; set; }
 }
