+++ src/Ray.BiliBiliTool.Agent/BiliBiliAgent/Dtos/Mall/Sign2Request.cs
@@ -2,13 +2,13 @@ namespace Ray.BiliBiliTool.Agent.BiliBiliAgent.Dtos.Mall;
 
 public class Sign2Request
 {
+    public string device { get; set; } = "phone";
+
     /// <summary>
     /// 当前时间（毫秒）
     /// </summary>
     /// <sample>1748445354567</sample>
     public long t { get; set; } = DateTimeOffset.UtcNow.ToUnixTimeMilliseconds();
-
-    public string device { get; set; } = "phone";
 
     /// <summary>
     /// 当前时间（秒）
