+++ src/Ray.BiliBiliTool.Agent/BiliBiliAgent/Dtos/VipTask/ThreeDaysSign/BigPointDto.cs
@@ -1,0 +1,3 @@
+namespace Ray.BiliBiliTool.Agent.BiliBiliAgent.Dtos.VipTask.ThreeDaysSign;
+
+public record BigPointDto(int point, int expire_point, int expire_time, int expire_days);
