+++ src/Ray.BiliBiliTool.Agent/BiliBiliAgent/Dtos/VipTask/ThreeDaysSign/ThreeDaySignRequest.cs
@@ -1,0 +1,6 @@
+namespace Ray.BiliBiliTool.Agent.BiliBiliAgent.Dtos.VipTask.ThreeDaysSign;
+
+public class ThreeDaySignRequest : BaseAppRequest
+{
+    public required string csrf { get; set; }
+}
